SaveSystem Load
El SaveSystem llegeix les dades guardades en el JSON i només guarda el nou temps si és menor que l'anterior.

+++ TFG/Assets/Scripts/SaveSystem.cs
@@ -7,17 +7,45 @@ public class SaveSystem
     public const string SaveFilePath = "/SaveFile.json";
 
     [Serializable]
-    private struct SaveFile
+    public struct SaveFile
     {
         public string time;
+        public float totalTime;
     }
 
-    public void Save(string time)
+    public void Save(string time, float totalTime)
     {
         SaveFile saveFile = new SaveFile();
         saveFile.time = time;
+        saveFile.totalTime = totalTime;
 
-        string json = JsonUtility.ToJson(saveFile);
-        File.WriteAllText(Application.dataPath + SaveFilePath, json);
+        if (File.Exists(Application.dataPath + SaveSystem.SaveFilePath))
+        {
+            SaveFile oldSaveFile = new SaveFile();
+            oldSaveFile = Load();
+
+            if(saveFile.totalTime < oldSaveFile.totalTime)
+            {
+                string json = JsonUtility.ToJson(saveFile);
+                File.WriteAllText(Application.dataPath + SaveFilePath, json);
+            }
+        }
+        else
+        {
+            
+            string json = JsonUtility.ToJson(saveFile);
+            File.WriteAllText(Application.dataPath + SaveFilePath, json);
+        }
+        
+    }
+
+    public SaveFile Load()
+    {
+        SaveFile saveFile = new SaveFile();
+        string json = File.ReadAllText(Application.dataPath + "/SaveFile.json");
+
+        saveFile = JsonUtility.FromJson<SaveFile>(json);
+
+        return saveFile;
     }
 }
